+++ Sources/Arithmetic.swift
@@ -1,0 +1,6 @@
+public protocol Arithmetic: Equatable {
+    static func + (_: Self, _: Self) -> Self
+    static func - (_: Self, _: Self) -> Self
+    static func * (_: Self, _: Self) -> Self
+    static func / (_: Self, _: Self) -> Self
+}