+++ Sources/Matrix_1x1.swift
@@ -1,7 +1,3 @@
-extension Matrix where M == _1, N == _1, Real == Double {
-    public var scalar: Real { return self[0, 0] }
-}
-
-extension Matrix where M == _1, N == _1, Real == Float {
-    public var scalar: Real { return self[0, 0] }
+extension Matrix where M == _1, N == _1 {
+    public var scalar: Field { return self[0, 0] }
 }