+++ Sources/Matrix_NxN.swift
@@ -1,15 +1,5 @@
-import Accelerate
-
-extension Matrix where M == N, Real == Double {
-    public static func identity() -> Matrix<M, N, Real> {
-        return Matrix(la_identity_matrix(N.value, la_scalar_type_t(LA_SCALAR_TYPE_DOUBLE), la_attribute_t(LA_DEFAULT_ATTRIBUTES)))
-    }
-}
-
-extension Matrix where M == N, Real == Float {
-    public static func identity() -> Matrix<M, N, Real> {
-        return Matrix(la_identity_matrix(N.value, la_scalar_type_t(LA_SCALAR_TYPE_FLOAT), la_attribute_t(LA_DEFAULT_ATTRIBUTES)))
-    }
+extension Matrix where M == N {
+    public static func identity() -> Matrix<M, N, Field> { return Field.matrix_identity() }
 }
 
 extension Matrix where M == N {