+++ CHANGELOG.md
@@ -1,8 +1,7 @@
 ## Unreleased
 
 - Rename `init` to `create`. It will conflict to other initializers in the future.
-- Support `Double` and `Float` as `Real` parameter.
-- `Matrix` doesn't conform `Equatable`.
+- Support `Double` and `Float` as `Field` parameter.
 
 
 ## 0.4.0